requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.5
flatbuffers==24.3.25
fonttools==4.55.1
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
html5lib==1.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
ml-dtypes==0.4.1
multitasking==0.0.11
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
protobuf==5.29.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yfinance==0.2.49
